Scale down textures, first pass at fixing skybox
Added code to scale textues down before loading them into video memory. My
graphics card can't handle TF2 itself at the highest quality, so this
should help. Planning to add UI to choose texture quality later.

Changed skybox code to cut the "side.png" texture in half, and use the
halves on alternating sides of the skybox. Not really sure how this should
be handled, and it's still not working completely, but this gets rid of
the non-square texture warning.