Only display the first submodel of each bodypart.
This fixes the heavy's discolored left hand. The heavy's hand bodypart has
three submodels: one with the correct appearance, one that doesn't appear
to have any mesh associated with it, and one that applies too much black
texture to the left hand. I was not able to find any other models with
multiple submodels out of the props used in ctf_2fort, so hopefully this
change won't have any side-effects.