Got the 2D skybox working.
As far as I can tell, the 2fort skybox textures only cover the top half of
the skybox, so I'm cropping and adjusting to satisfy that.

The 3D skybox is not being displayed at all, so some black will be visible
below the horizon where that would go. Presumably the 3D skybox would need
a separate render pass, after the normal part of the map, but before the
skybox and transparent models.