Figured out how to extract lightmaps from BSP.
So this is just another preprocessing step, but unzipping inside the
browser would be pointless without writing code to convert VTF files
inside the browser. The downside is that this information will be
downloaded twice, once in the second half (yes, half) of the BSP, and once
as separate files.

Fun fact: It's also possible to extract these files by identifying the
first PK\x03\x04 magic number in the BSP, chopping off that part of the
file, and then treating it as a standard .zip file.